Implement ImageModal component for photo viewing and navigation
