+++ src/routes/photos/+page.svelte
@@ -3,6 +3,7 @@
 	import { goto, afterNavigate } from '$app/navigation';
 	import { ImageViewer } from 'svelte-image-viewer';
 	import { page } from '$app/state';
+	import ImageModal from './ImageModal.svelte';
 
 	interface Photo {
 		src: string;
@@ -23,7 +24,6 @@
 	let modalOpen = false;
 	let modalPhoto: Photo | null = null;
 	let modalIndex = 0;
-	let modalContainer: HTMLDivElement | null = null;
 	let touchStartX: number | null = null;
 	let zoomed = false;
 
@@ -191,11 +191,6 @@
 	onDestroy(() => {
 		document.body.style.overflow = '';
 	});
-
-	// Focus modal for keyboard navigation
-	$: if (modalOpen && modalContainer) {
-		setTimeout(() => modalContainer && modalContainer.focus(), 0);
-	}
 </script>
 
 <!-- Breadcrumbs -->
@@ -265,262 +260,16 @@
 	<p class="text-center text-gray-500">No photos found in this section.</p>
 {/if}
 
-<!-- Photo Modal -->
-{#if modalOpen && modalPhoto}
-	<div
-		bind:this={modalContainer}
-		class="fixed inset-0 z-50 flex items-center justify-center bg-black/90"
-		role="dialog"
-		aria-modal="true"
-		tabindex="0"
-		on:keydown={(e) => {
-			if (!zoomed) {
-				if (e.key === 'ArrowLeft') showPrev();
-				if (e.key === 'ArrowRight') showNext();
-			}
-			if (e.key === 'Escape') {
-				if (zoomed) zoomed = false;
-				else closeModal();
-			}
-			setTimeout(() => modalContainer && modalContainer.focus(), 0);
-		}}
-		on:click={() => setTimeout(() => modalContainer && modalContainer.focus(), 0)}
-	>
-		<!-- Overlay for closing modal -->
-		<button
-			type="button"
-			aria-label="Close photo modal"
-			class="absolute inset-0"
-			on:click={closeModal}
-			tabindex="-1"
-			style="background: transparent; border: none; padding: 0; margin: 0;"
-		></button>
-
-		<!-- Modal content -->
-		<div
-			class="relative z-10 flex h-screen max-h-none w-screen max-w-none flex-col bg-transparent"
-			role="presentation"
-			on:click|self={() => setTimeout(() => modalContainer && modalContainer.focus(), 0)}
-			on:click|stopPropagation
-		>
-			<!-- Top bar -->
-			<div class="flex w-full items-center justify-between px-6 pt-6 pb-2">
-				<!-- Counter styled like arrow buttons -->
-				{#if currentSection}
-					<div class="flex items-center rounded-full bg-black/60 px-4 py-2 text-xs text-white">
-						<svg class="mr-1 h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor"
-							><circle cx="12" cy="12" r="10" stroke-width="2" /></svg
-						>
-						<span>{modalIndex + 1} / {currentSection.photos.length}</span>
-					</div>
-				{/if}
-				<div class="flex items-center gap-2">
-					<!-- Zoom icon styled like arrow buttons -->
-					<button
-						class="flex items-center justify-center rounded-full bg-black/60 p-3 text-white transition hover:bg-gray-700"
-						on:click={() => {
-							zoomed = !zoomed;
-							if (zoomed) {
-								// Optionally, trigger zoom-to-1:1 if your ImageViewer supports it
-							}
-						}}
-						aria-label={zoomed ? 'Disable zoom/pan' : 'Enable zoom/pan'}
-					>
-						{#if zoomed}
-							<!-- Magnifier minus SVG -->
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								class="h-6 w-6"
-								fill="none"
-								viewBox="0 0 24 24"
-								stroke="currentColor"
-								><circle cx="11" cy="11" r="7" stroke-width="2" /><path
-									stroke-width="2"
-									d="M21 21l-4.35-4.35"
-								/><path stroke-width="2" d="M8 11h6" /></svg
-							>
-						{:else}
-							<!-- Magnifier plus SVG -->
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								class="h-6 w-6"
-								fill="none"
-								viewBox="0 0 24 24"
-								stroke="currentColor"
-								><circle cx="11" cy="11" r="7" stroke-width="2" /><path
-									stroke-width="2"
-									d="M21 21l-4.35-4.35"
-								/><path stroke-width="2" d="M11 8v6M8 11h6" /></svg
-							>
-						{/if}
-					</button>
-					<!-- Close icon styled like arrow buttons -->
-					<button
-						class="flex items-center justify-center rounded-full bg-black/60 p-3 text-xl text-white transition hover:bg-gray-700"
-						on:click={closeModal}
-						aria-label="Close"
-					>
-						<svg
-							xmlns="http://www.w3.org/2000/svg"
-							class="h-6 w-6"
-							fill="none"
-							viewBox="0 0 24 24"
-							stroke="currentColor"
-							><path stroke-linecap="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" /></svg
-						>
-					</button>
-				</div>
-			</div>
-
-			<!-- Image area -->
-			<div
-				class="group relative flex flex-1 items-center justify-center select-none"
-				style="min-height:0;"
-				aria-label="Next or previous photo"
-				role="button"
-				tabindex="0"
-				on:click={(e) => {
-					if (e.target instanceof HTMLElement && !e.target.closest('button')) {
-						if (!zoomed) {
-							const bounds = e.currentTarget.getBoundingClientRect();
-							const x = e.clientX - bounds.left;
-							if (x < bounds.width / 2) showPrev();
-							else showNext();
-						}
-						requestAnimationFrame(() => modalContainer && modalContainer.focus());
-					}
-				}}
-				on:keydown={(e) => {
-					if (e.target !== e.currentTarget) return; // Only handle if this div is focused
-					if (zoomed) return;
-					if (e.key === 'Enter' || e.key === ' ') {
-						// Default to next photo on Enter/Space
-						showNext();
-						requestAnimationFrame(() => modalContainer && modalContainer.focus());
-						e.preventDefault();
-					}
-				}}
-				on:touchstart={handleTouchStart}
-				on:touchend={handleTouchEnd}
-			>
-				<!-- Left arrow (always visible on desktop) -->
-				{#if !zoomed}
-					<button
-						class="absolute top-1/2 left-4 z-20 -translate-y-1/2 rounded-full bg-black/20 p-3 text-white md:flex"
-						on:click|stopPropagation={() => {
-							showPrev();
-							requestAnimationFrame(() => modalContainer && modalContainer.focus());
-						}}
-						tabindex="-1"
-						aria-label="Previous photo"
-					>
-						<svg
-							xmlns="http://www.w3.org/2000/svg"
-							class="h-7 w-7"
-							fill="none"
-							viewBox="0 0 24 24"
-							stroke="currentColor"
-							><path
-								stroke-linecap="round"
-								stroke-linejoin="round"
-								stroke-width="2"
-								d="M15 19l-7-7 7-7"
-							/></svg
-						>
-					</button>
-				{/if}
-
-				<!-- ImageViewer or plain img depending on zoomed state -->
-				{#if zoomed}
-					<ImageViewer src={modalPhoto.src} alt={modalPhoto.title} />
-				{:else}
-					<img
-						src={modalPhoto.src}
-						alt={modalPhoto.title}
-						class="max-h-full max-w-full object-contain select-none"
-						draggable="false"
-						tabindex="-1"
-					/>
-				{/if}
-
-				<!-- Right arrow (always visible on desktop) -->
-				{#if !zoomed}
-					<button
-						class="absolute top-1/2 right-4 z-20 -translate-y-1/2 rounded-full bg-black/20 p-3 text-white md:flex"
-						on:click|stopPropagation={() => {
-							showNext();
-							requestAnimationFrame(() => modalContainer && modalContainer.focus());
-						}}
-						tabindex="-1"
-						aria-label="Next photo"
-					>
-						<svg
-							xmlns="http://www.w3.org/2000/svg"
-							class="h-7 w-7"
-							fill="none"
-							viewBox="0 0 24 24"
-							stroke="currentColor"
-							><path
-								stroke-linecap="round"
-								stroke-linejoin="round"
-								stroke-width="2"
-								d="M9 5l7 7-7 7"
-							/></svg
-						>
-					</button>
-				{/if}
-			</div>
-
-			<!-- EXIF/Title block (below image) -->
-			<div class="w-full bg-black/70 p-4 text-white dark:bg-black/80">
-				<div class="text-lg font-bold">{modalPhoto.title}</div>
-				<div class="text-sm">{modalPhoto.date}</div>
-				<div class="mt-1 text-xs">
-					{modalPhoto.camera}
-					{modalPhoto.lens}
-					{modalPhoto.focalLength ? ` | ${modalPhoto.focalLength}` : ''}
-					{modalPhoto.aperture ? ` | ${modalPhoto.aperture}` : ''}
-					{modalPhoto.exposure ? ` | ${modalPhoto.exposure}` : ''}
-					{modalPhoto.iso ? ` | ISO ${modalPhoto.iso}` : ''}
-				</div>
-			</div>
-
-			<!-- Filmstrip pinned to bottom of modal content -->
-			{#if currentSection && currentSection.photos.length > 1}
-				<div
-					class="flex w-full items-center justify-center gap-2 overflow-x-auto bg-black/70 px-4 py-2 dark:bg-black/80"
-					style="z-index:20;"
-				>
-					{#each currentSection.photos as thumb, idx}
-						<button
-							type="button"
-							class="mx-1 cursor-pointer rounded border-2 transition-all"
-							style="border-color: {idx === modalIndex
-								? 'var(--color-accent)'
-								: '#444'}; outline: none;"
-							on:click={(e) => {
-								e.preventDefault();
-								if (!currentSection) return;
-								modalIndex = idx;
-								modalPhoto = currentSection.photos[idx];
-								const params = new URLSearchParams(page.url.search);
-								params.set('photo', idx.toString());
-								goto(`${window.location.pathname}?${params.toString()}`, { replaceState: true });
-								requestAnimationFrame(() => modalContainer && modalContainer.focus());
-							}}
-							tabindex="-1"
-							aria-label={`Go to photo ${idx + 1}`}
-						>
-							<img
-								src={thumb.src}
-								alt={thumb.title}
-								class="h-12 w-auto rounded object-cover"
-								style="opacity: {idx === modalIndex ? 1 : 0.6}; border-radius: 4px;"
-							/>
-						</button>
-					{/each}
-				</div>
-			{/if}
-		</div>
-	</div>
-{/if}
+<ImageModal
+	open={modalOpen}
+	photos={currentSection?.photos ?? []}
+	index={modalIndex}
+	section={currentSection}
+	on:close={closeModal}
+	on:change={(e) => {
+		modalIndex = e.detail.index;
+		if (currentSection) {
+			modalPhoto = currentSection.photos[modalIndex];
+		}
+	}}
+/>
